+++ packages/widget/src/components/SelectTokenButton/SelectTokenButton.tsx
@@ -8,6 +8,7 @@ import { useWidgetConfig } from '../../providers/WidgetProvider/WidgetProvider.j
 import type { FormTypeProps } from '../../stores/form/types.js'
 import { FormKeyHelper } from '../../stores/form/types.js'
 import { useFieldValues } from '../../stores/form/useFieldValues.js'
+import { HiddenUI } from '../../types/widget.js'
 import { navigationRoutes } from '../../utils/navigationRoutes.js'
 import { AvatarBadgedDefault, AvatarBadgedSkeleton } from '../Avatar/Avatar.js'
 import { TokenAvatar } from '../Avatar/TokenAvatar.js'
@@ -25,7 +26,7 @@ export const SelectTokenButton: React.FC<
 > = ({ formType, compact }) => {
   const { t } = useTranslation()
   const navigate = useNavigate()
-  const { disabledUI, subvariant } = useWidgetConfig()
+  const { disabledUI, subvariant, hiddenUI } = useWidgetConfig()
   const swapOnly = useSwapOnly()
   const tokenKey = FormKeyHelper.getTokenKey(formType)
   const [chainId, tokenAddress] = useFieldValues(
@@ -50,7 +51,8 @@ export const SelectTokenButton: React.FC<
   const defaultPlaceholder =
     formType === 'to' && subvariant === 'refuel'
       ? t('main.selectChain')
-      : formType === 'to' && swapOnly
+      : (formType === 'to' && swapOnly) ||
+          hiddenUI?.includes(HiddenUI.ChainSelect)
         ? t('main.selectToken')
         : t('main.selectChainAndToken')
   const cardTitle: string =